+++ test/dummy/app/models/page.rb
@@ -7,6 +7,7 @@ class Page < ApplicationRecord
     root: false,
     allowed_parent_types: ["Workspace", "Folder"]
   )
+  RecordingStudio.enable_capability(:accessible, on: self)
 
   validates :title, presence: true
 end